fix(diff): handle YAML short-forms like '!GetAtt' in diff (#10381)
CloudFormation allows using short-form versions of intrinsic functions like `!GetAtt`.
We handled them correctly in the `@aws-cdk/cloudformation-include` module,
so extract that logic to a common package,
and use it from the CLI in the `diff` command as well.

Fixes #6537

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*
