+++ package.json
@@ -61,14 +61,14 @@
       "@aws-cdk/cloud-assembly-schema/jsonschema/**",
       "@aws-cdk/cloud-assembly-schema/semver",
       "@aws-cdk/cloud-assembly-schema/semver/**",
-      "@aws-cdk/cloudformation-include/yaml",
-      "@aws-cdk/cloudformation-include/yaml/**",
       "@aws-cdk/core/fs-extra",
       "@aws-cdk/core/fs-extra/**",
       "@aws-cdk/core/minimatch",
       "@aws-cdk/core/minimatch/**",
       "@aws-cdk/cx-api/semver",
       "@aws-cdk/cx-api/semver/**",
+      "@aws-cdk/yaml-cfn/yaml",
+      "@aws-cdk/yaml-cfn/yaml/**",
       "aws-cdk-lib/case",
       "aws-cdk-lib/case/**",
       "aws-cdk-lib/fs-extra",
