+++ packages/@aws-cdk/cloudformation-include/package.json
@@ -68,6 +68,7 @@
     "@aws-cdk/aws-apigatewayv2": "0.0.0",
     "@aws-cdk/aws-appconfig": "0.0.0",
     "@aws-cdk/aws-applicationautoscaling": "0.0.0",
+    "@aws-cdk/aws-applicationinsights": "0.0.0",
     "@aws-cdk/aws-appmesh": "0.0.0",
     "@aws-cdk/aws-appstream": "0.0.0",
     "@aws-cdk/aws-appsync": "0.0.0",
@@ -89,6 +90,7 @@
     "@aws-cdk/aws-codecommit": "0.0.0",
     "@aws-cdk/aws-codedeploy": "0.0.0",
     "@aws-cdk/aws-codeguruprofiler": "0.0.0",
+    "@aws-cdk/aws-codegurureviewer": "0.0.0",
     "@aws-cdk/aws-codepipeline": "0.0.0",
     "@aws-cdk/aws-codestar": "0.0.0",
     "@aws-cdk/aws-codestarconnections": "0.0.0",
@@ -180,9 +182,7 @@
     "@aws-cdk/aws-wafv2": "0.0.0",
     "@aws-cdk/aws-workspaces": "0.0.0",
     "@aws-cdk/core": "0.0.0",
-    "yaml": "1.10.0",
-    "@aws-cdk/aws-applicationinsights": "0.0.0",
-    "@aws-cdk/aws-codegurureviewer": "0.0.0"
+    "@aws-cdk/yaml-cfn": "0.0.0"
   },
   "peerDependencies": {
     "@aws-cdk/alexa-ask": "0.0.0",
@@ -194,6 +194,7 @@
     "@aws-cdk/aws-apigatewayv2": "0.0.0",
     "@aws-cdk/aws-appconfig": "0.0.0",
     "@aws-cdk/aws-applicationautoscaling": "0.0.0",
+    "@aws-cdk/aws-applicationinsights": "0.0.0",
     "@aws-cdk/aws-appmesh": "0.0.0",
     "@aws-cdk/aws-appstream": "0.0.0",
     "@aws-cdk/aws-appsync": "0.0.0",
@@ -215,6 +216,7 @@
     "@aws-cdk/aws-codecommit": "0.0.0",
     "@aws-cdk/aws-codedeploy": "0.0.0",
     "@aws-cdk/aws-codeguruprofiler": "0.0.0",
+    "@aws-cdk/aws-codegurureviewer": "0.0.0",
     "@aws-cdk/aws-codepipeline": "0.0.0",
     "@aws-cdk/aws-codestar": "0.0.0",
     "@aws-cdk/aws-codestarconnections": "0.0.0",
@@ -257,6 +259,7 @@
     "@aws-cdk/aws-iotanalytics": "0.0.0",
     "@aws-cdk/aws-iotevents": "0.0.0",
     "@aws-cdk/aws-iotthingsgraph": "0.0.0",
+    "@aws-cdk/aws-kendra": "0.0.0",
     "@aws-cdk/aws-kinesis": "0.0.0",
     "@aws-cdk/aws-kinesisanalytics": "0.0.0",
     "@aws-cdk/aws-kinesisfirehose": "0.0.0",
@@ -296,6 +299,7 @@
     "@aws-cdk/aws-sns": "0.0.0",
     "@aws-cdk/aws-sqs": "0.0.0",
     "@aws-cdk/aws-ssm": "0.0.0",
+    "@aws-cdk/aws-sso": "0.0.0",
     "@aws-cdk/aws-stepfunctions": "0.0.0",
     "@aws-cdk/aws-synthetics": "0.0.0",
     "@aws-cdk/aws-transfer": "0.0.0",
@@ -304,25 +308,18 @@
     "@aws-cdk/aws-wafv2": "0.0.0",
     "@aws-cdk/aws-workspaces": "0.0.0",
     "@aws-cdk/core": "0.0.0",
-    "constructs": "^3.0.4",
-    "@aws-cdk/aws-applicationinsights": "0.0.0",
-    "@aws-cdk/aws-codegurureviewer": "0.0.0",
-    "@aws-cdk/aws-kendra": "0.0.0",
-    "@aws-cdk/aws-sso": "0.0.0"
+    "@aws-cdk/yaml-cfn": "0.0.0",
+    "constructs": "^3.0.4"
   },
   "devDependencies": {
     "@aws-cdk/assert": "0.0.0",
     "@types/jest": "^26.0.14",
-    "@types/yaml": "1.9.6",
     "cdk-build-tools": "0.0.0",
     "cdk-integ-tools": "0.0.0",
     "jest": "^26.4.2",
     "pkglint": "0.0.0",
     "ts-jest": "^26.3.0"
   },
-  "bundledDependencies": [
-    "yaml"
-  ],
   "keywords": [
     "aws",
     "cdk",
