+++ packages/@aws-cdk/yaml-cfn/.eslintrc.js
@@ -1,0 +1,3 @@
+const baseConfig = require('cdk-build-tools/config/eslintrc');
+baseConfig.parserOptions.project = __dirname + '/tsconfig.json';
+module.exports = baseConfig;
