+++ packages/@aws-cdk/yaml-cfn/.gitignore
@@ -1,0 +1,18 @@
+*.js
+*.js.map
+*.d.ts
+node_modules
+dist
+tsconfig.json
+.jsii
+
+.LAST_BUILD
+.LAST_PACKAGE
+*.snk
+.nyc_output
+coverage
+nyc.config.js
+!.eslintrc.js
+!jest.config.js
+
+junit.xml
