+++ packages/@aws-cdk/yaml-cfn/NOTICE
@@ -1,0 +1,2 @@
+AWS Cloud Development Kit (AWS CDK)
+Copyright 2018-2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
