+++ packages/@aws-cdk/yaml-cfn/README.md
@@ -1,0 +1,16 @@
+## CloudFormation YAML utilities
+
+<!--BEGIN STABILITY BANNER-->
+---
+
+![cdk-constructs: Experimental](https://img.shields.io/badge/cdk--constructs-experimental-important.svg?style=for-the-badge)
+
+> The APIs of higher level constructs in this module are experimental and under active development. They are subject to non-backward compatible changes or removal in any future version. These are not subject to the [Semantic Versioning](https://semver.org/) model and breaking changes will be announced in the release notes. This means that while you may use them, you may need to update your source code when upgrading to a newer version of this package.
+
+---
+<!--END STABILITY BANNER-->
+
+This module contains utilities for parsing and emitting
+YAML that is used by [AWS CloudFormation](https://aws.amazon.com/cloudformation).
+
+This module is part of the [AWS Cloud Development Kit](https://github.com/aws/aws-cdk) project.
