+++ packages/@aws-cdk/yaml-cfn/jest.config.js
@@ -1,0 +1,4 @@
+const baseConfig = require('cdk-build-tools/config/jest.config');
+module.exports = {
+    ...baseConfig,
+};
