+++ packages/@aws-cdk/yaml-cfn/lib/index.ts
@@ -1,0 +1,1 @@
+export * from './yaml';
