+++ packages/@aws-cdk/yaml-cfn/test/yaml.test.ts
@@ -1,0 +1,5 @@
+import '@aws-cdk/assert/jest';
+
+test('No tests are specified for this package', () => {
+  expect(true).toBe(true);
+});
