+++ packages/aws-cdk-lib/package.json
@@ -253,6 +253,7 @@
     "@aws-cdk/cx-api": "0.0.0",
     "@aws-cdk/pipelines": "0.0.0",
     "@aws-cdk/region-info": "0.0.0",
+    "@aws-cdk/yaml-cfn": "0.0.0",
     "@types/fs-extra": "^8.1.1",
     "@types/node": "^10.17.35",
     "cdk-build-tools": "0.0.0",
