+++ packages/decdk/package.json
@@ -183,6 +183,7 @@
     "@aws-cdk/cx-api": "0.0.0",
     "@aws-cdk/pipelines": "0.0.0",
     "@aws-cdk/region-info": "0.0.0",
+    "@aws-cdk/yaml-cfn": "0.0.0",
     "constructs": "^3.0.4",
     "fs-extra": "^9.0.1",
     "jsii-reflect": "^1.12.0",
